requirements.txt
flask
requests
websockets
waitress
python-dotenv
